Handle issuers with non-standard paths
An issuer _should_ use the same base path as the corresponding
path in the namespace.  This is not always true, however.  The
CHTC issuer, for example, uses the base path of /chtc for the namespace
/chtc/PROTECTED.

The detection logic must know this in order to determine if a given
token is applicable.