Properly (re-)initialize config subsystem during the tests
In some unit tests, we need to alter the configuration via env var
(to ensure the environment-based changes are effective); this adds
in re-initialization at the appropriate points.

Additionally, this moves the various timeout constants to the config
subsystem, allowing us to use much shorter timeouts for test runs
(taking about 6 minutes off the time to run `go test`).