HTCONDOR-1731 Try replacing dots in token names with underscores
While adding support for underscores in URL schemes was laudible,
there is still a problem parsing the (illegal) URLs upstream in
HTCondor. Until we come up with a permanent solution, our workaround
is to have users submit with a "." in place of the "_" (similar to
how the client handles underscores internally) and have any plugins
do the conversion from "." to "_" when doing token discovery.

Here I have refactored the HTCondor-specific token discovery code
into its own function, which will search for both tokens with
and without "." replaced with "_" in their filenames. The
TestGetToken test was expanded to test URL schemes using both "_"
and "." and token files named with both "_" and ".".